fix: move test types to root 'type' folder and remove 'typeRoots' property from tsconfig
