--- test/types/index.d.ts
@@ -1,13 +1,0 @@
-export {} //indicate it is a module type declaration
-
-declare global {
-  interface Window {
-    BeeJs: typeof import('../../src')
-  }
-  namespace jest {
-    interface Matchers<R> {
-      toBeHashReference(): R
-      toBeBeeResponse(statusCode: number): R
-    }
-  }
-}
