+++ tsconfig.json
@@ -10,8 +10,7 @@
     "strict": true,
     "resolveJsonModule": true,
     "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
-    "typeRoots": ["./src/types"]
+    "emitDecoratorMetadata": true
   },
   "exclude": ["dist", "node_modules"]
 }
