+++ types/index.d.ts
@@ -1,2 +1,13 @@
-import * as TestTypes from '../test/types'
-export { TestTypes }
+export {} //indicate it is a module type declaration
+
+declare global {
+  interface Window {
+    BeeJs: typeof import('../src')
+  }
+  namespace jest {
+    interface Matchers<R> {
+      toBeHashReference(): R
+      toBeBeeResponse(statusCode: number): R
+    }
+  }
+}
